+++ src/floating-select-infinite/index.js
@@ -1,0 +1,1 @@
+export FloatingSelectInfinite from './floating-select-infinite.component';