+++ src/floating-select-windowed/index.js
@@ -1,1 +1,0 @@
-export FloatingSelectWindowed from './floating-select-windowed.component';